fix: workflow job list pagination to handle large job counts (#478)
* fix: workflow job list pagination to handle large job counts

Prior to this commit, obtaining the job ID via gh CLI would fail
when in a workflow with a job count that exceeds 100.
For example, in a workflow with 105 jobs, the last 5 jobs fail
to obtain their ids, because the 100-sized page didn't include them.

This PR fixes the issue by using the `--paginate` flag to list
_all_ of the jobs even when they exceed 100 jobs. It uses the
built-in pagination feature of the CLI that allows for using
multiple pages under the hood when necessary, while still surfacing
the output JSON as a single list (not split by pages).

This allows for workflows with arbitrarily-many jobs,
at least up to the theoretical limits of how many JSON bytes
we can pass around as a string inside our bash commands here.

* replace `per_page=100` with `paginate`

---------

Co-authored-by: Rishav Dhar <19497993+rdhar@users.noreply.github.com>